example/requirements.txt
Flask-SQLAlchemy==2.0
Flask==0.10.1
multiget-cache==0.0.8
./../